Made getEndpointById from test-helper actually work on the server
